requirements.txt
opencv-python
numpy
matplotlib
scipy
scikit-image
PySimpleGUI
Pillow
scikit-fuzzy